cloud_functions/slackbot_consume/requirements.txt
aiohttp==3.6.2
ansible==2.8.5
async-timeout==3.0.1
asyncio==3.4.3
attrs==19.3.0
beautifulsoup4==4.9.1
cachetools==4.1.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cloudevents==0.3.0
cryptography==2.9.2
flake8==3.8.3
Flask==1.1.2
functions-framework==2.0.0
google-api-core==1.22.0
google-api-python-client==1.10.0
google-auth==1.20.0
google-auth-httplib2==0.0.4
google-cloud-core==1.3.0
google-cloud-pubsub==1.7.0
google-cloud-storage==1.30.0
google-crc32c==0.1.0
google-resumable-media==0.7.0
googleapis-common-protos==1.52.0
grpc-google-iam-v1==0.12.3
grpcio==1.30.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.9
importlib-metadata==1.6.1
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.5.1
MarkupSafe==1.1.1
mccabe==0.6.1
multidict==4.7.6
pathtools==0.1.2
protobuf==3.12.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pyee==7.0.2
pyflakes==2.2.0
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
requests==2.23.0
rq==1.4.2
rsa==4.6
six==1.14.0
slackclient==2.7.3
slackeventsapi==2.1.0
soupsieve==2.0.1
tableauserverclient==0.12
uritemplate==3.0.1
urllib3==1.25.9
watchdog==0.10.3
websocket-client==0.54.0
Werkzeug==1.0.1
yarl==1.4.2
zipp==3.1.0